Added getToken.js
