+++ sdk/requests/constants.js
@@ -1,0 +1,7 @@
+const REQUEST_TYPES = {
+  LOGIN: 'login',
+  GET_TOKEN: 'getToken',
+  GET_REFRESH_TOKEN: 'getRefreshToken',
+};
+
+export default REQUEST_TYPES;
